requirements.txt
click==6.7
dominate==2.3.1
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Session==0.3.1
Flask-SocketIO==2.9.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.1
python-engineio==1.7.0
python-socketio==1.7.7
scikits.audiolab==0.11.0
scipy==0.19.1
six==1.10.0
visitor==0.1.3
Werkzeug==0.12.2
